# Conflicts:
#	src/Views/EstatusViaje.js
#	src/Views/Guia.js
#	src/Views/Informes.js
#	src/Views/Tarifas/Tarifas.js
#	src/Views/TipoCobro.js
#	src/Views/Viajes.js

+++ src/Views/Guia.js
@@ -388,7 +388,7 @@ function Guia(props) {
             console.log(err)
             showSuccess(err)
         });
-    
+
     }
 
     function handleEliminar(id) {
@@ -401,18 +401,35 @@ function Guia(props) {
                 showSuccess("El usuario no tiene derechos para realizar el proceso");
                 return;
             }
-            validarEliminarGuia(id).then(respuesta => {
-                if (respuesta.data.sePuedeEliminar) {
-                    eliminarGuia(id, state.modificadoPor).then(respuesta => {
-                        showSuccess(respuesta.data)
-                        if (respuesta.data.indexOf("fracaso:") <= 0)
-                            getAllData()
-                        setGuiaSeleccionada(null)
-                    }).catch(err => {
-                        console.log(err)
-                        showSuccess(err.response?.data)
-                    });
-                } else {
+            validarEliminarGuia(id).then(respuesta=>{
+                if(respuesta.data.sePuedeEliminar){
+
+                    confirmAlert({
+                        title: 'Confirmar Eliminar',
+                        message: '¿Está seguro de eliminar guia?',
+                        buttons: [
+                            {
+                                label: 'Si',
+                                onClick: () => {
+                                    eliminarGuia(id, state.modificadoPor).then(respuesta => {
+                                        showSuccess(respuesta.data)
+                                        if (respuesta.data.indexOf("fracaso:") <= 0) {
+                                            getAllData()
+                                            setGuiaSeleccionada(null)
+                                        }
+                                    }).catch(err => {
+                                        console.log(err)
+                                        showSuccess(err.response?.data)
+                                    });
+                                }
+                            },
+                            {
+                                label: 'No',
+                            }
+                        ]
+                    })
+
+                }else{
                     showSuccess("La guia no puede ser eliminada a menos que se cancele")
                 }
             })
